extract constants to `constants.js`. add `isSketchStringsEqual` function to resolve issues with sketch "strings"
