+++ Contents/Sketch/constants.js
@@ -1,0 +1,16 @@
+const SYMBOL_INSTANCE_CLASS_NAME = "MSSymbolInstance";
+const SYMBOL_MASTER_CLASS_NAME = "MSSymbolMaster";
+
+const LEGEND_ARTBOARD_NAME = 'Legend';
+const LEGEND_ITEM_INDEX_NAME = 'LegendItemIndex';
+const LEGEND_ARTBOARD_MIN_WIDTH = 300;
+const LEGEND_ITEM_HEIGHT = 80;
+
+module.exports = {
+  SYMBOL_INSTANCE_CLASS_NAME,
+  SYMBOL_MASTER_CLASS_NAME,
+  LEGEND_ARTBOARD_NAME,
+  LEGEND_ITEM_INDEX_NAME,
+  LEGEND_ARTBOARD_MIN_WIDTH,
+  LEGEND_ITEM_HEIGHT,
+};
